+++ pygal/state.py
@@ -30,6 +30,7 @@ class State(object):
 
     def __init__(self, graph, **kwargs):
         """Create the transient state"""
+        self.__dict__.update(**graph.config.__class__.__dict__)
         self.__dict__.update(**graph.config.__dict__)
         self.__dict__.update(**graph.__dict__)
         self.__dict__.update(**kwargs)